refactor(SLB-256): expose operation helper types from the codegen plugin
