+++ packages/npm/@amazeelabs/codegen-operation-ids/tsconfig.json
@@ -10,6 +10,7 @@
     "resolveJsonModule": false,
     "jsx": "react"
   },
+  "include": ["src"],
   "$schema": "https://json.schemastore.org/tsconfig",
   "display": "Recommended"
 }
